+++ continuum-core/src/main/java/org/apache/maven/continuum/core/action/AddBuildDefinitionToProjectGroupAction.java
@@ -42,7 +42,7 @@ public class AddBuildDefinitionToProjectGroupAction
         throws Exception
     {
         BuildDefinition buildDefinition = getBuildDefinition( map );
-        int projectGroupId =  getProjectGroupId( map );
+        long projectGroupId =  getProjectGroupId( map );
 
         ProjectGroup projectGroup = store.getProjectGroupWithBuildDetails( projectGroupId );
